+++ src/main/java/com/project/banking/service/BankService.java
@@ -1,7 +1,6 @@
 package com.project.banking.service;
 
 import com.project.banking.model.Bank;
-import com.project.banking.model.BankCurrency;
 import com.project.banking.response.ApiResponse;
 import com.project.banking.response.BankCurrencyResponse;
 
@@ -11,7 +10,7 @@ import java.util.List;
 public interface BankService {
 
     ApiResponse createBank(Bank bank, List<Integer> bankCurrencies);
-    ApiResponse updateBank(Integer id, float amount);
+    ApiResponse updateBank(Integer id, double amount);
     ApiResponse deleteBank(Integer id);
     Bank getBankById(Integer id);
     Bank getBankByName(String name);
